refactor: split generic CheckCollection content into new check_collections subpackage
Move CheckCollection* base classes plus the generic dataclasses, enums,
helper classes, and protocols out of soda_core/contracts/ into a new
soda_core/check_collections/ subpackage.

The Contract* subclasses stay in their existing soda_core/contracts/
locations as thin BC shim subclasses. The original modules add re-exports
of moved symbols so existing imports of the form
'from soda_core.contracts.contract_verification import Contract' keep
working unchanged.

New files:
- soda_core/check_collections/__init__.py
- soda_core/check_collections/check_collection_verification.py
- soda_core/check_collections/impl/__init__.py
- soda_core/check_collections/impl/check_collection_yaml.py
- soda_core/check_collections/impl/check_collection_verification_impl.py

No behavior change. All public import paths preserved via re-exports.

Refs DES-315
